backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-dotenv==1.0.1
motor==3.3.2
pymongo==4.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.26.0
google-auth==2.28.1
google-auth-oauthlib==1.2.0
google-api-python-client==2.118.0
twilio==8.13.0
scikit-learn==1.4.0
numpy==1.26.4
Pillow==10.2.0
deepface==0.0.89
fer==22.5.1
opencv-python-headless==4.9.0.80
pydantic==2.6.1
pydantic-settings==2.2.1
aiohttp==3.9.3
python-multipart==0.0.9
cryptography==42.0.3
